[MCS RE] fixes cyclomatic complexity
